+++ scanapi/evaluators/code_evaluator.py
@@ -20,7 +20,7 @@ class CodeEvaluator:
 
     @classmethod
     def evaluate(cls, sequence, vars, is_a_test_case=False):
-        match = cls.python_code_pattern.search(sequence)
+        match = cls.python_code_pattern.search(str(sequence))
 
         if not match:
             return sequence