+++ tests/unit/evaluators/test_code_evaluator.py
@@ -9,7 +9,7 @@ from scanapi.evaluators import CodeEvaluator, SpecEvaluator, StringEvaluator
 class TestCodeEvaluator:
     class TestEvaluate:
         class TestWhenDoesNotMatchThePattern:
-            test_data = ["no code", "${CODE}", "${code}", "{{code}}"]
+            test_data = ["no code", "${CODE}", "${code}", "{{code}}", 10, []]
 
             @pytest.mark.parametrize("sequence", test_data)
             def test_should_return_sequence(self, sequence):