+++ tests/unit/evaluators/test_string_evaluator.py
@@ -145,6 +145,8 @@ class TestStringEvaluator:
                 "https://jsonplaceholder.typicode.com",
                 "https://jsonplaceholder.typicode.com/posts",
             ),
+            ("${product_id}", "${product_id}", 100, 100),
+            ("products/${product_id}", "${product_id}", 100, "products/100"),
         ]
 
         @pytest.mark.parametrize(